+++ .prettierrc
@@ -1,5 +1,5 @@
 {
-  "printWidth": 80,
+  "printWidth": 120,
   "trailingComma": "all",
   "singleQuote": true
 }
