+++ src/__tests__/strict-http-firewall.tests.ts
@@ -1,5 +1,4 @@
-import { HttpFirewallOptions } from '../types';
-import { StrictHttpFirewall } from '../strict.http.firewall';
+import { httpFirewall, HttpFirewallOptions } from '../strict-http-firewall';
 import express from 'express';
 import request from 'supertest';
 
@@ -10,10 +9,8 @@ describe('HttpStrictFirewall test suite', () => {
       const options: HttpFirewallOptions = {
         allowedHttpMethods: ['POST', 'GET'],
       };
-      app.use(new StrictHttpFirewall(options).firewall);
-      const res = await request(app)
-        .head('/')
-        .set('Content-Type', 'application/json');
+      app.use(httpFirewall(options));
+      const res = await request(app).head('/').set('Content-Type', 'application/json');
       expect(res.statusCode).toBe(403);
     });
 
@@ -23,16 +20,14 @@ describe('HttpStrictFirewall test suite', () => {
       const options: HttpFirewallOptions = {
         allowedHttpMethods: ['POST', 'GET'],
       };
-      app.use(new StrictHttpFirewall(options).firewall);
+      app.use(httpFirewall(options));
       app.get('/', (req, res) => {
         // You're working with an express req and res now.
         res.status(200).send();
       });
 
       // Act
-      const res = await request(app)
-        .get('/')
-        .set('Content-Type', 'application/json');
+      const res = await request(app).get('/').set('Content-Type', 'application/json');
 
       // Assert
       expect(res.statusCode).toBe(200);
