+++ src/demo/demo.ts
@@ -1,15 +1,11 @@
 import express, { Express, Request, Response } from 'express';
-import { HttpFirewallOptions, Predicate } from '../types';
-import { StrictHttpFirewall } from '../index';
+import { httpFirewall, HttpFirewallOptions, Predicate } from '../index';
 
 const app: Express = express();
 const port = 5428;
 
 // This must be added first, before adding any routes
-app.use(new StrictHttpFirewall(firewallOptions()).firewall);
-
-// Or, to simply use the firewall with default rules:
-//app.use(httpFirewall)
+app.use(httpFirewall(firewallOptions()));
 
 app.get('/', (req: Request, res: Response) => {
   res.send('Http Firewall Demo running');
@@ -21,9 +17,9 @@ app.listen(port, () => {
 
 function firewallOptions(): HttpFirewallOptions {
   // Allows traffic from specific hosts only
-  const allowedHostnamesPredicate = Predicate.of<string>((h) =>
-    h.endsWith('example.com'),
-  ).or(Predicate.of<string>((h) => h === 'localhost'));
+  const allowedHostnamesPredicate = Predicate.of<string>((h) => h.endsWith('example.com')).or(
+    Predicate.of<string>((h) => h === 'localhost'),
+  );
 
   return {
     allowedHostnames: allowedHostnamesPredicate,
