+++ src/index.ts
@@ -1,1 +1,6 @@
-export * from './strict.http.firewall';
+export {
+  httpFirewall,
+  HttpFirewallOptions,
+  HttpMethod,
+  Predicate
+} from './strict-http-firewall';
